Moving main code into subpackage prior to chopping it up.
